refactor(project): move template symbol identity
