+++ crates/djls-project/src/lib.rs
@@ -62,12 +62,12 @@ pub use specs::KnownOptions;
 pub use specs::ModelGraph;
 pub use specs::RequiredKeyword;
 pub use specs::SplitPosition;
-pub use specs::SymbolKey;
+pub use templates::SymbolKey;
 pub use specs::TagArgument;
 pub use specs::TagArgumentKind;
 pub use specs::TagRule;
 pub use specs::TagRuleMap;
-pub use specs::TemplateSymbolKind;
+pub use templates::TemplateSymbolKind;
 pub use specs::compute_model_graph;
 pub use specs::extract_block_specs;
 pub use specs::extract_filter_arities;
