+++ crates/djls-project/src/specs.rs
@@ -18,6 +18,7 @@ use ruff_python_ast::StmtFunctionDef;
 use crate::ast::Recurse;
 use crate::ast::walk_stmts;
 use crate::templates::RegistrationInfo;
+use crate::templates::SymbolKey;
 use crate::templates::collect_registrations_from_body;
 use crate::names::ModulePath;
 use crate::parse::parse_python_module;
@@ -44,12 +45,10 @@ pub use crate::specs::types::FilterArityMap;
 pub use crate::specs::types::KnownOptions;
 pub use crate::specs::types::RequiredKeyword;
 pub use crate::specs::types::SplitPosition;
-pub use crate::specs::types::SymbolKey;
 pub use crate::specs::types::TagArgument;
 pub use crate::specs::types::TagArgumentKind;
 pub use crate::specs::types::TagRule;
 pub use crate::specs::types::TagRuleMap;
-pub use crate::specs::types::TemplateSymbolKind;
 
 /// Interned key for a helper function call.
 ///
